src/requirements.txt
astroid==3.0.1
colorama==0.4.6
configargparse==1.7
dill==0.3.7
imageio==2.31.1
imutils==0.5.4
isort==5.12.0
lazy-loader==0.3
mccabe==0.7.0
networkx==3.1
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pillow==10.0.0
platformdirs==4.0.0
pylint==3.0.2
pyperclip==1.8.2
pytesseract==0.3.10
pywavelets==1.4.1
scikit-image==0.21.0
scipy==1.9.3
tifffile==2023.7.18
tomlkit==0.12.3
